Reorganise modules to make them more cohesive (#41)
This PR reorganizes substantial parts of the former domain Maven module,
now called core.

It introduces two new root packages: domain and application. The domain model knowledge is now structured in the first, and everything that belongs to application logic and domain object orchestration is in the ladder.

With this we hope to improve the class organization and reflect the actual domain and application structure.