Re-arrange architectural composition (#69)
Changes the data manager architecture into components (represented by Maven modules) and makes use of Java modules to enforce stronger encapsulation of the components' implementation detail.

We draft four two modules (BLUE), that will contain domain logic: authentication and authorization.

In addition, we introduce a module for broadcasting information through the application but decouple it from the domains as good as possible and still allow for gossip.

The newsreader module holds the current logic of consuming messages from the broadcasting module by implementing the MessageSubscriber interface of the broadcasting module. When required later, this can be divided into own sub-components and realize own (micro)services if required.