Introduce form input validation strategy (#199)
Uses Vaadin's Binder approach to validate input fields of forms.

Exemplary, the approach is shown for the project information dialog.

In addition, we introduce a showcase for registering to component events that
are important for parent pages to understand the state of the user flow. If a user 
indicates to cancel a form submission in a dialog, the parent components needs to know. 

It is bad if we expose the detailed components to the outside (like the buttons), so we use ComponentEvents instead. Same is true for FormSubmissions, where we use semantically
sound event names, like ProjectCreationEvent.