Fix array out of bound exception
An exception was thrown whenever a list of values was pasted into a cell for special cases.
Whenever the resulting values exceeded the rows or columns present in the spreadsheet, the index was out of bounds for the internal lists of column or rows.
By wrapping the result into an optional and checking the indices beforehand, this can be avoided.