Remove cell comments and always close dialogue to address remaining spreadsheet issues (#451)
* Remove comment creation for error messages

* Remove comment css and checks

* Always close dialog independent if batch creation/edit failed or succeeded

* Ensure the dialog is closed before the notifications are generated

* Handle error output from sample registration service

* Refactor closing of event source in SampleContentComponent

The closing of the event source in the SampleContentComponent is moved to a different part in the process of registering and editing a batch. Specifically, it now closes after a successful action or when an error occurs. This change ensures the correct execution of the application even in case of failures during batch modifications.

* fix validation check

* fix tests

---------

Co-authored-by: Tobias Koch <tobias.koch@qbic.uni-tuebingen.de>
Co-authored-by: Tobias Koch <kochtobi@users.noreply.github.com>