Remove biological replicate from concept (#582)
We came to the conclusion together with some stakeholders, that the concept of the biological replicate
is too complex and also technically hard to solve. It leads to a drop in UX, since it makes the sample
registration process more complex and it is hard to find human readable representations of biological id references.

The user is able to provide an organism ID for a sample, so we leave it to the user to use custom labels to identify true biological replicates for
any given condition.