Separate Spring Datasources (#576)
* separate datasources

* separate databases

* Introduce separate environment variables for the datasources

* AclSecurity uses IdentityDataSource

* Remove overwriting `@EntityScan`

The `@EntityScan` annotation probably caused the repositories to already be there when the datasources tried to instantiate them. Thus the default datasource was used for all repositories.

* Remove overwriting `@EntityScan`

The `@EntityScan` annotation probably caused the repositories to already be there when the datasources tried to instantiate them. Thus the default datasource was used for all repositories.

* add java doc

* correct links to documentation

* add hibernate ddl settings

* move authorization to project management

* remove default property

* move sql files

* format sql queries

* restructure datasources