Fix broken ProjectAccessComponent
Due to the way the grid constructor initializes the grid, the grid is now generated with columns for each bean property.
As we add the columns ourselves, this leads to every column already being present (duplicate column key) and the ProjectAccessComponent.java not being instantiated.
Using the no-args constructor of the Grid class solves this issue.