Fix redundant ontology terms that are actually the same (#932)
Due to the improvement of the OntologyTerm comparison by considering equal terms when they have the same IRI, the combobox showed empty rows for a term that was referenced in more than two ontologies.

This contribution filters them out.

Also, an unused ontology term list is removed. The NCBI Taxon filtering happens in the SpeciesSearchService and should also not be knowledgable by the client.