Introduce stepper dialog  (#934)
A stepper dialog (also known as wizard), contains at least two steps that are displayed consecutively to the user and ask them for input. They are usually applied when the user input is more complex, is conditionally chained, or can be separated contextually into logical steps.

This contribution introduces a stepper dialog component, that is easy to build and extend, and simplifies the stepper dialog creation.