Implement reactive API for sample validation (#1093)
Implements the API for sample validation and wires the new service in the UI as well.

This commit also introduces the concepts required for measurement validation as well as the service implementation.

The wiring with the UI for measurements is more complex due to the design of the dialog. The dialog should be refactored and then the service wiring adjusted, which will then be straight forward.