Introduces optimistic locking for project aggregate (#1127)
Concurrent modification of projects is now possible due to the introduction
of an optimistic locking strategy introducing a version for every row entry
for the project aggregate.

In case of a concurrent write modification, an exception is thrown and
the retry handler of the flux jumps in with the default retry strategy we have
currently configured.