Redo metadata conversion logic (#1168)
Improves the metadata conversion logic and introduces
a converter registry for different metadata conversion scenarios.

The service API is refactored to allow for a more
expressive handling of pooled measurements to prepare the collation
of pooled measurements on the IO layer already and not 
consider that as service logic anymore.