Add Client ID for ROR repository
Due to a change coming december, the ROR repository requires the addition of a client id to the http request. This PR introduces these changes. It does not adapt the data manager to being able to use the `v2` api of the ROR repository.
For more information see https://ror.readme.io/docs/client-id.