Improve cache utilisation on term lookup (#1246)
Inverts the order of cache and remote service lookup for OBO ID resolve. Now, the cache is queried first and then the remote service is requested.
This leads to a significant lower remote service request load and minimises to risk to receive a GO_AWAY from the remote server.
Also the OBO ID Lookup is faster, since on batch measurement registrations on the same machine, worst case is that the ID need to be resolved once if not in cache.

Also improves logging for easier debugging.


Co-authored-by: Tobias Koch <tobias.koch@uni-tuebingen.de>

* Lower log level to debug for cache lookup failure

* Provide OBO ID in error log message

---------

Co-authored-by: Tobias Koch <tobias.koch@uni-tuebingen.de>