Fix display of orcid account on Chrome browser (#1305)
* Replace width 100% with flex grow

Firefox and chrome interpret 100% inside a flex layout differently. Chrome tries to assign the `spacer` a width of 100% by shrinking the public record component. Leading to a screwed display.
By telling the browsers that the spacer should expand instead of that the spacer should fill the full width, both firefox and chrome display the component correctly.

* Change CSS class name

Co-authored-by: Sven F. <sven1103@users.noreply.github.com>

---------

Co-authored-by: KochTobi <kochtobi@users.noreply.github.com>
Co-authored-by: Sven F. <sven1103@users.noreply.github.com>