Introduce signature check of dependencies (#1307)
Introduces the maven pgp verify plugin and activates it for every module that has a src/main directory.

Builds will now fail, if the signature does not validate and there is no rule specified for it in the pgp-keys.override.list.

To skip manual curation of known maven artifacts, we use a pgp-keys-map from org.simplify4u.