Only query for ontology classes (#1315)
Without a type definition all types of `class,property,individual,ontology` are returned. As `property` does not provide a parseable curie/obo_id, our integration fails for every query where a property or ontology is returned. By specifying the type to be a `class`, we get only expected results.

Documentation on the API:
https://api.terminology.tib.eu/swagger-ui/index.html?urls.primaryName=version1

Co-authored-by: KochTobi <kochtobi@users.noreply.github.com>