Fix UI freeze during measurement validation and registration (#1319)
Removes some smelling code that can produce race conditions for the UI update. Also passes now the current UI instance as reference for background jobs with reactor. Ui getters are not thread-safe and are not guaranteed to return the intended UI reference. According to Vaadin's documentation, it is better to fetch the UI reference in the main thread and pass it to background threads explicitly.

Also, this PR enforces the server to push changes to the UI, since with automatic polling we see that the UI is not always updated despite this should be the expected behaviour when the user session lock is released. This resulted in ui components not being updated from background tasks (validation, registration pending task toast, etc) and the update was waiting for a ui interaction from the user.