Fix wrong push mode (#1335)
There is no need to set the push mode to Manual. If there would be a need to do so it should be reset to the original value as not to impact other views. As it is not necessary, I left it at AUTOMATIC instead. This resolves https://github.com/qbicsoftware/data-manager-app/issues/1317

Co-authored-by: KochTobi <kochtobi@users.noreply.github.com>