Update version during release for all project modules (#1348)
Fixes the CI build.
As the project structure changed and the bom module does not have the datamanager module as a parent, we need to modify the version bumping. In addition to that the snapshots are now deployed when PRs against main are opened (ncluding hotfix prs so they can be tested before merging)
distributionManagement is provided for the bom module as it is not inherited.